Create README.md
A novel method for presenting crime and safety data to the college public. While the Clery Act mandates that a daily crime log be maintained, not all constituencies know about the log or are able to easily glean relevant information from it. Our solution leverages online analytical processing (OLAP) to answer  multi-dimensional analytical queries about crime and safety in and around the Claremont Colleges as they are posed by members of the public.